+++ src/prop/g12/common/StubGraf.java
@@ -6,43 +6,43 @@ import java.util.Arrays;
 
 public class StubGraf<T> {
 	
-	Integer[][] adjacencia;
+	Double[][] adjacencia;
 	Integer numNodes;
 	ArrayList<T> nodes;
 	int c;
 
 	public StubGraf() {
 		numNodes = 5;
-		adjacencia = new Integer[5][5];
-		adjacencia[0][0] = -1;
-		adjacencia[0][1] = 0;
-		adjacencia[0][2] = -1;
-		adjacencia[0][3] = 2;
-		adjacencia[0][4] = -1;
+		adjacencia = new Double[5][5];
+		adjacencia[0][0] = -1.0;
+		adjacencia[0][1] = 0.0;
+		adjacencia[0][2] = -1.0;
+		adjacencia[0][3] = 2.0;
+		adjacencia[0][4] = -1.0;
 		
-		adjacencia[1][0] = 0;
-		adjacencia[1][1] = -1;
-		adjacencia[1][2] = 1;
-		adjacencia[1][3] = 1;
-		adjacencia[1][4] = -1;
+		adjacencia[1][0] = 0.0;
+		adjacencia[1][1] = -1.0;
+		adjacencia[1][2] = 1.0;
+		adjacencia[1][3] = 1.0;
+		adjacencia[1][4] = -1.0;
 		
-		adjacencia[2][0] = -1;
-		adjacencia[2][1] = 1;
-		adjacencia[2][2] = -1;
-		adjacencia[2][3] = -1;
-		adjacencia[2][4] = 1;
+		adjacencia[2][0] = -1.0;
+		adjacencia[2][1] = 1.0;
+		adjacencia[2][2] = -1.0;
+		adjacencia[2][3] = -1.0;
+		adjacencia[2][4] = 1.0;
 		
-		adjacencia[3][0] = 2;
-		adjacencia[3][1] = 1;
-		adjacencia[3][2] = -1;
-		adjacencia[3][3] = -1;
-		adjacencia[3][4] = 1;
+		adjacencia[3][0] = 2.0;
+		adjacencia[3][1] = 1.0;
+		adjacencia[3][2] = -1.0;
+		adjacencia[3][3] = -1.0;
+		adjacencia[3][4] = 1.0;
 		
-		adjacencia[4][0] = -1;
-		adjacencia[4][1] = -1;
-		adjacencia[4][2] = 1;
-		adjacencia[4][3] = 1;
-		adjacencia[4][4] = -1;
+		adjacencia[4][0] = -1.0;
+		adjacencia[4][1] = -1.0;
+		adjacencia[4][2] = 1.0;
+		adjacencia[4][3] = 1.0;
+		adjacencia[4][4] = -1.0;
 		
 		nodes = new ArrayList<T> (5);
 		c = 5;
@@ -60,11 +60,11 @@ public class StubGraf<T> {
 		return nodes;
 	}
 	
-	public Integer GetAresta(int idNode1, int idNode2) {
+	public Double GetAresta(int idNode1, int idNode2) {
 		return adjacencia[idNode1][idNode2];
 	}
 	
-	public Integer[][] GetGraf() {
+	public Double[][] GetGraf() {
 		return adjacencia;
 	}
